KSR-385 Table tab save when tab is closed
